Refactor home page
