+++ src/assets/icons.tsx
@@ -1,0 +1,7 @@
+import { FaCode, FaDatabase, FaTools } from 'react-icons/fa'
+
+export const FrontendIcon: React.FC = () => <FaCode className="text-blue-600 dark:text-blue-300 w-12 h-12" />
+
+export const BackendIcon: React.FC = () => <FaDatabase className="text-green-600 dark:text-green-300 w-12 h-12" />
+
+export const DevOpsIcon: React.FC = () => <FaTools className="text-yellow-600 dark:text-yellow-300 w-12 h-12" />
