+++ src/components/index.ts
@@ -1,0 +1,5 @@
+import Hero from './Hero'
+import SkillCard from './SkillCard'
+import ProjectCard from './ProjectCard'
+
+export { Hero, SkillCard, ProjectCard }
