+++ src/data/index.ts
@@ -1,0 +1,4 @@
+import { projects } from './projects'
+import { skills } from './skills'
+
+export { projects, skills }
