+++ src/data/projects.ts
@@ -1,0 +1,20 @@
+export const projects = [
+  {
+    title: 'Personal Portfolio',
+    description: 'A responsive and dynamic portfolio to showcase my work, built with Next.js and TailwindCSS.',
+    link: 'https://github.com/vladyslav-holovashchenko-web/portfolio',
+    tags: ['Next.js', 'TailwindCSS', 'React'],
+  },
+  {
+    title: 'E-Commerce Platform',
+    description: 'A full-stack application with real-time product search, cart functionality, and user authentication.',
+    link: 'https://github.com/vladyslav-holovashchenko-web/ecommerce',
+    tags: ['Node.js', 'MongoDB', 'React'],
+  },
+  {
+    title: 'Chat Application',
+    description: 'A real-time chat application with socket.io for instant messaging and an intuitive user interface.',
+    link: 'https://github.com/vladyslav-holovashchenko-web/chat-app',
+    tags: ['Socket.io', 'Node.js', 'Express'],
+  },
+]
