+++ src/data/skills.tsx
@@ -1,0 +1,19 @@
+import { FrontendIcon, BackendIcon, DevOpsIcon } from '@/assets/icons'
+
+export const skills = [
+  {
+    title: 'Frontend Development',
+    description: 'Expertise in React, Next.js, TailwindCSS, and creating responsive, accessible UIs.',
+    icon: <FrontendIcon />,
+  },
+  {
+    title: 'Backend Development',
+    description: 'Proficient in Node.js, Express, and database technologies like MongoDB and PostgreSQL.',
+    icon: <BackendIcon />,
+  },
+  {
+    title: 'DevOps & Tools',
+    description: 'Skilled in Docker, Git, CI/CD pipelines, and cloud platforms like AWS and Vercel.',
+    icon: <DevOpsIcon />,
+  },
+]
